api/hum_service/pyproject.toml
[tool.poetry]
name = "hum_service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.75.1"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
python-multipart = "^0.0.5"
python-magic = "^0.4.25"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
python-magic-bin = "^0.4.14"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
